fix perf: memory issues after changing CHR palette many times
Hard to give an estimate, but suppose 50 frames per double click, limit was about 9600 or so, so around 192 updates. The fix works by subscribing a lambda instead of a bound method. For whatever reason, tag bind doesn't 'attach' to funcs, but it does to objects. So something to do with the types most likely, or maybe something about extra references for objects or whatever. But it works fine for funcs, just as long as you don't ever ever send bound methods or attributes. So no self at all. Instead, we send createanims, which is a parameter here, same for index, and then with that we're able to index into the right object and call the corresponding method. It's so beautiful. Well for PalRectangles is a key but still indexing of sorts. And yeah unbinding didn't work either, del didn't work, pretty much nothing worked until using functions instead. So awesome.