fix+refactor: some TileImages are still not deleted completely
I mean not deleted at all. I mean to say, the deletion is not complete as a whole. You double right click, tag_bind happens but it was still attaching to the object, no lambda. So now it is. And refactor, cause I moved it to a single point of entrance :)