fix: error when clicking on Stop Anim
Due to a race condition of sorts. Messagebox stops flow, but play_anim_label was already destroyed. The anim keeps playing because while the label has been destroyed, the flag hasn't been set to False yet. So it tries to place the label, but the label was already destroyed. Error. Solution, don't show the messagebox if we're loading because of a stop anim.