feat: add physics to play anim
Looks so good! I will decide on a future commit on initial x and y values. I was thinking, maybe for some anims I can open up a new window, but perhaps I can instead give an option of starting at a slightly different x and y, or well a completely different ones, or maybe I can adjust a little bit the height... I will see on that, but this will do now.