fix: after pal update and hover old value is still shown
That's the effect, yes. It's one of those bugs where you can describe them in two different ways. The should way, like, well if I'm updating the color of the PalRectangle object, then its corresponding attribute should be updated too, right? So we could say that's the bug. But then you don't see much in the front end that you could consider an issue or a bug from user perspective. It's kinda like, it should but should it? I mean if it's not impacting anything. But in this case, it is. So, we can discuss what description to use but in this case I went for the user perspective description, and then here in the details I explain the technicalities of it. Yeah maybe you could say any bug is about shoulds but I think there are more notorious shoulds and other less notorious? Like, there are things that you look and hey, that should be updated, hey, that should be cleared. Like, according to coding styles and whatever anyways, I do agree that it's a very... kind of gray area, so I think that's also part of why I'm sticking to the UI perspective, that's more objective.