refactor: animify palette updates
I know, animify, yes! Like, now palettes work pretty much the same way as Anim and CHR. I also updated the dimensions of the canvas to also make it look more consistent. This is very much necessary for UndoRedo to fully work (it was breaking pretty bad if you changed different character_palette indexes, if you tried to undo after a palette refresh etc. etc. and maaaaaybe it was possible with the previous code but, it didn't feel like I wanted to keep going that route. This does it, works like a charm, works great!