feat: add traceback in case of any errors
YES! Found a way! Also, I can compile (yes, compile, like whatever) using --noconsole option and it works so beautifully well!

Amend: but to clarify, hard crashes will still not be logged! The fail to allocate bitmap is presumably a message that is printed to the console by Tkinter before something else happens that causes the Access Violation error that finally causes the hard crash.