feat: add CreateAnimsUpdater
So we could check for updates before, we would display messageboxes, we could run create_release for a version, we could restore commit hash to previous value so that it isn't commited too but still included in executable, so Check for Updates itself was working, here we're adding the download and installation. But, there's no UI for now. That's in next commit.