feat: add try-except blocks for CreateAnimsUpdater
CreateAnimsUpdater will rarely update. Even when performing an update, only the executable will be updated, well and potentially the README.html, everything else won't, I mean just the updater won't. So from this version on, I'd like to have a lot of try-except what is also called error handling and add also that if there's an error, maybe it's because we do need to download directly from the repo. Alright!