feat: add clear all selections
Oh yes! The time did come!

Amend: Celebrate! And, yeah doing an amend because I did want to say celebrate and because I'm adding the pal_label update as well. I was going to consider it a fix, but you wouldn't notice until this change, this feature. You Ctrl+D, for example, and now you click on the palette, but the palette label is not updated. Now it is.