feat: add loading bar
Yay! Cleanup is coming in a later commit, maybe the next one? I'm also including docs here and not sure if you'd call it a refactor? So I didn't add it in commit title but, I moved Tk() creation to init_anim_window which made more sense and, in practice, it didn't change anything so it's not tied to this change but... whatever.