fix!: width and height's value refs the UndoRedo
Resulting in all sorts of chaos. Oh my goodness, I'm still so happy, exhausted, anxious, delighted, a bit of everything, that I found out that. I know I know, I'm giving myself lots of flowers but I mean this makes me so happy. On to the next features! Oh wait, no explanation? I mean it's in the comments. But summary, on its own didn't change anything, that's why this time I'm not describing the bug in terms of impact because there is none, at least not that I was able to find. But, there is an impact in the future, in the next commit. After adding pasting CHR rectangles, if you update width, undo, paste, switch branch, redo, and undo, you'll see corrupted tiles, more specifically, you'll see the tiles from the previous redo, which sounds absolutely crazy and insane until you understand exactly what's going on. I could have rubbed it off as well not that important it still works but it was way too strange that it had different behavior with new_tile_indexes, I'm so glad I checked this one out.