refactor: several rectangle updates in Anims and TileUtils
I think rectangle is a good word. Refactors like use a tag instead of a flag which can have None or the id of the rectangle, use select instead of calculating x and y and recreating rectangles again, don't use delete_tile_image_rectangles, use same method Canvas.delete with tag, update naming so that it's more similar between Anim and TileUtils and there are still the differences that have to be there like with pasting and whatnot (naming I mean like tile_selected_object instead of tile_image_object, because it's not about the image itself, it's about the selected), remove some checks that are not necessary anymore with the refactors. I mean before the delete_tile_image_rectangles would give an error because we'd be passing None instead of the IDs. But now we don't do that anymore, we delete with the tag.